feat(FE-45): unify admin and user registration in AuthenticationRegister
- Unified admin and user registration into a single reusable AuthenticationRegister component.
- Introduced `isAdminRegister` prop to conditionally render admin-specific fields.
- Added support for role selection and user assignment (students/teachers) when registering as admin.
- Preserved all validation and UI behavior for both flows.
- Removed the now-redundant AdminRegister component to simplify maintenance.
