+++ front-end/src/components/admin/registerNewUsers/RegisterNewUsers.component.tsx
@@ -13,12 +13,11 @@ const RegisterNewUser: React.FC<IRegisterNewUserProps> = ({
   const { classes } = useStyles();
   return (
     <AuthenticationRegister
+      isAdminRegister={true}
       rootClassName={classes.root}
       adminRefetchUserList={refetchUserList}
       registerTitle={'Register new users'}
-      hideSocialButtons={false}
       showNotification={true}
-      isAdminRegister={true}
     />
   );
 };
