+++ front-end/src/components/navBar/AdminTabs.component.tsx
@@ -3,14 +3,12 @@ import {
   IconPencil,
   IconUserCog,
   IconUserPlus,
-  IconSchool,
 } from '@tabler/icons-react';
 import { useAppDispatch } from '../../context/AppContext';
 import React from 'react';
 const data = [
   { link: 'userManagment', label: 'User Management', icon: IconUserCog },
   { link: 'register_new_user', label: 'Register New User', icon: IconUserPlus },
-  { link: 'RegisterUser', label: 'RegisterUser', icon: IconSchool },
   { link: 'Assignent', label: 'Assign assigment', icon: IconPencil },
   { link: 'TestIt', label: 'Testme', icon: IconPencil },
 ];
