+++ front-end/src/components/navBar/deleteAfter/AdminNavBar.component.tsx
@@ -6,7 +6,6 @@ import {
   IconLogout,
   IconUserCog,
   IconUserPlus,
-  IconSchool,
   IconPencil,
 } from '@tabler/icons-react';
 import React from 'react';
@@ -16,7 +15,6 @@ import { useAppDispatch, useAppState } from '../../../context/AppContext';
 const data = [
   { link: 'userManagment', label: 'User Management', icon: IconUserCog },
   { link: 'register_new_user', label: 'Register New User', icon: IconUserPlus },
-  { link: 'RegisterUser', label: 'RegisterUser', icon: IconSchool },
   { link: 'Assignent', label: 'Assign Assignment', icon: IconPencil },
   { link: 'TestIt', label: 'Test Me', icon: IconPencil },
 ];
